+++ activesupport/test/ordered_hash_test.rb
@@ -1,6 +1,6 @@
 require 'abstract_unit'
 require 'active_support/json'
-require 'active_support/core_ext/object/to_json'
+require 'active_support/core_ext/object/json'
 require 'active_support/core_ext/hash/indifferent_access'
 require 'active_support/core_ext/array/extract_options'
